Suggest wrapping Collapse child in no-margin node
It makes a big difference when you have elements with margin or padding;
the animation appears much smoother when the contents are enclosed
in a node with zero padding and zero margin (like a raw `<div>`).

Here are two examples.
  * with a div: https://cloud.githubusercontent.com/assets/4317806/10871987/4bf4ea48-80ac-11e5-84f5-45c955a22f47.gif
  * without a div: https://cloud.githubusercontent.com/assets/4317806/10871988/504aee62-80ac-11e5-9232-7dc97c151fef.gif
(source: https://github.com/WChargin/lc3/commit/344da3aa497204722c9a8c47dccc37a984e5d4e4)

The example already wraps its contents in a `<div>`,
but it's not clear that or why this is beneficial.