[changed] Make Panel expansion cancelable
Pass event to PanelGroup onSelect handler. This changes the public API for the onSelect handler from `function handleSelect(eventKey)` to `function handleSelect(eventKey, event)`.

Because the onSelect handler receives the DOM event, it can check what the target element was. When combined with `event.selected = false`, clicks on specific elements can be ignored, e.g. a button with onClick logic is included in the panel and clicking it should not toggle the panel.