fix: add missing props to Navbar.Collapse (#3519)
I tried testing this change in a dependent project, but installing the package from local clone just results in `Cannot find module 'react-bootstrap/Navbar'`. To complicate matters, the build setup seems closely tied to the contributors' development environment. The second commit in this PR fixes the build from a clean state; it only ever worked from a dirty state. There is no `prepare` NPM script: is that intentional?

At the least, this can serve as a starting point.

/cc @jquense