Add "as" prop to Form.Check
It can be useful to override the element/component used as the
underlying form component. This was already possible in Form.Control via
its "as" prop but not in Form.Check.

Example use case: the Formik library provides form components which pull
form context information and automatically control form element state
and event handlers. It's possible to use its component for checkboxes or
radio buttons by passing its Field component with this "as" prop to
Form.Check.

Closes #4565