fix(carousel): re-implement with committed ref
Re-implements the reference to the interval value with a committed
ref, ensuring that the interval value we're accessing is safe
when accessing the value inside of effects.

This in turn fixes the consistency issues within our unit tests
(although I still ended up increasing the time being ticked to
give a little more leeway for the events being fired).

This also includes a re-work of traversing the children to grab
the interval values at the same time that we count the children,
removing that potential side effect from the rendering logic.